requirements.txt
aiogram==3.4.1
aiogram-calendar==1.0.0
python-dotenv==1.0.1
asyncpg==0.29.0
matplotlib==3.8.4
openpyxl==3.1.2
